requirements.txt
google-cloud-bigquery>=3.4.0
google-cloud-storage>=2.7.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
scipy>=1.9.0
faiss-cpu>=1.7.0
tqdm>=4.64.0